Function props typing proposal
Hi. Thanks a lot for the post. I enjoyed it.

I would like to send you a proposal about, due to the same typing structure for `Home`, `Page` and `Navigation` components are repeated in every one, what about if that structure is defined by an interface and then, this interface is reused where it's needed?

This way the code becomes more readable and its maintenance is a little bit easier because every component is typed from a single source of truth.

Thanks a lot for the time you spent on reading my proposal.