Adding a suggestion (#641)
This is just a suggestions. I'm sure the same can be better written and perhaps that's not necessarily the right place in the article for it.
I just found out this the other day. Our file had 100% coverage, yet there was no test for the case where the condition was false. This caused a bug in production.