Update index.mdx
i'm not sure about the wording, but as far as I understand this, the value can't be something which is included in the FalsyType, as the FalsyTypes are excluded from the ValueType. So the value is everything but the FalsyTypes.

Additional thought: the typedBoolean function returns true or false not based on the types. The function itself is merely a help for a compiler to narrow down the value types which will be passed to the function. Thats why the whole sentence is a bit misleading.